The localtime function takes a time_t* argument.
Passing the ut_time member of the utmp struct as this
argument doesn't work on 64 bit systems because the lengths
of the arguments don't agree.  So use a temproary variable
to store the ut_time value in a time_t variable and pass
the address of the time_t variable to localtime.